+++ R/NMorderColumns.R
@@ -28,27 +28,30 @@
 ##'     last. NA's must be NA or ".".
 ##' @param alpha Sort columns alphabetically. Notice, this is the last
 ##'     order priority applied.
+##' @param col.id Name of the (numeric) unique subject ID. Can be
+##'     controlled with `NMdataConf()`.
 ##' @param col.nomtime The name of the column containing nominal
 ##'     time. If given, it will put the column quite far left, just
-##'     after row counter and ID. Default value is NOMTIME and can be
-##'     configured with NMdataConf.
+##'     after row counter and `col.id`. Default value is NOMTIME and can be
+##'     configured with `NMdataConf()`.
 ##' @param col.time The name of the column containing actual time. If
 ##'     given, it will put the column quite far left, just after row
-##'     counter, subject ID, and nominal time. Default value is TIME.
+##'     counter, subject ID, and nominal time. Default value is `TIME`. Can be
+##'     controlled with `NMdataConf()`.
 ##' @param col.row A row counter column. This will be the first column
 ##'     in the dataset. Technically, you can use it for whatever
-##'     column you want first. Default value is ROW and can be
-##'     configured with NMdataConf.
+##'     column you want first. Default value is `ROW` and can be
+##'     configured with `NMdataConf()`.
 ##' @param col.flagn The name of the column containing numerical flag
 ##'     values for data row omission. Default value is FLAG and can be
-##'     configured with NMdataConf.
+##'     configured with `NMdataConf()`.
 ##' @param col.dv a vector of column names to put early to represent
 ##'     dependent variable(s). Default is DV.
 ##' @param as.fun The default is to return a data.table if data is a
 ##'     data.table and return a data.frame in all other cases. Pass a
 ##'     function in as.fun to convert to something else. The default
-##'     can be configured using NMdataConf. However, if data is a
-##'     data.table, settings via NMdataConf are ignored.
+##'     can be configured using `NMdataConf()`. However, if data is a
+##'     data.table, settings via `NMdataConf()` are ignored.
 ##' @param quiet If true, no warning will be given about missing
 ##'     standard Nonmem columns.
 ##' @details This function will change the order of columns but it
@@ -82,8 +85,9 @@ NMorderColumns <- function(data,
                            lower.last=FALSE,
                            chars.last=TRUE,
                            alpha=TRUE,
+                           col.id,
                            col.nomtime,
-                           col.time="TIME",
+                           col.time,
                            col.row,
                            col.flagn,
                            col.dv="DV",
@@ -103,11 +107,12 @@ NMorderColumns <- function(data,
 ### Section end: Dummy variables, only not to get NOTE's in pacakge checks
     if(missing(col.flagn)) col.flagn <- NULL
     col.flagn <- NMdataDecideOption("col.flagn",col.flagn)
+    if(missing(col.id)) col.id <- NULL
+    col.id <- NMdataDecideOption("col.id",col.id)
     if(missing(col.nomtime)) col.nomtime <- NULL
     col.nomtime <- NMdataDecideOption("col.nomtime",col.nomtime)
-    ## col.time can currently not be specified with NMdataConf
-    ## if(missing(col.time)) col.time <- NULL
-    ## col.time <- NMdataDecideOption("col.time",col.time)
+    if(missing(col.time)) col.time <- NULL
+    col.time <- NMdataDecideOption("col.time",col.time)
     if(missing(quiet)) quiet <- NULL
     quiet <- NMdataDecideOption("quiet",quiet)
     
@@ -124,7 +129,7 @@ NMorderColumns <- function(data,
     if(missing(first)) first <- NULL
     if(missing(last)) last <- NULL
     
-    first1 <- c(col.row,"ID",col.nomtime,col.time,"EVID","CMT","AMT","II","ADDL","RATE",
+    first1 <- c(col.row,col.id,col.nomtime,col.time,"EVID","CMT","AMT","II","ADDL","RATE",
                 "SS",col.dv,"MDV")
     first2 <- c(col.flagn,"OCC","GRP","TRIAL","STUDY","DRUG","ROUTE")
 
@@ -132,7 +137,7 @@ NMorderColumns <- function(data,
     nms <- names(data)
     nms.dup <- nms[duplicated(nms)]
     if(!quiet && length(nms.dup)) messageWrap(paste0("Duplicated column names:\n",paste(nms.dup,collapse=", ")),fun.msg=warning)
-    ### these checks are now done in NMcheckData
+### these checks are now done in NMcheckData
     ## missing <- setdiff(setdiff(first1,c("II","ADDL","RATE","SS")),nms)
     ## if(!quiet && length(missing)) messageWrap(paste0("These standard Nonmem columns were not found in data:\n",paste(missing,collapse="\n")),fun.msg=message)